Record pressed precondition keys across modes
Used for hint2-2-then-click-mode.to.hint1-mode=^{rightalt} +backspace
rightalt was pressed before (in a previous mode) and its press state would not be recorded. hint2-2-then-click-mode would think that rightalt was not pressed even though it was, and the combo was completed even though it should have not been.